Source_code.cpp
I just write all functions in only one .cpp file, this is a little stupid but the program works well. The main idea of this code is to initialize the scheduling of all the task will running in the machine and then order them in order to save energy and time as much as we can.